+++ src/Database/InfluxDB/Manage.hs
@@ -1,17 +1,35 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE ViewPatterns #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+#if __GLASGOW_HASKELL__ >= 800
+{-# OPTIONS_GHC -Wno-missing-signatures #-}
+#else
+{-# OPTIONS_GHC -fno-warn-missing-signatures #-}
+#endif
 module Database.InfluxDB.Manage
-  ( manage
+  ( -- * Management query interface
+    Query
+  , manage
 
+  -- * Query parameters
+  , QueryParams
+  , queryParams
+  , server
+  , database
+  , precision
+  , manager
+
+  -- * Management query results
+  -- ** SHOW QUERIES
   , ShowQuery
   , qid
   , queryText
-  , Types.database
   , duration
 
+  -- ** SHOW SERIES
   , ShowSeries
   , key
   ) where
@@ -30,22 +48,32 @@ import qualified Data.Attoparsec.Combinator as AC
 import qualified Data.Attoparsec.Text as AT
 import qualified Data.Text.Encoding as TE
 import qualified Data.Vector as V
+import qualified Network.HTTP.Client as HC
 import qualified Network.HTTP.Types as HT
 
+import Database.InfluxDB.JSON (getField, parseQueryField)
 import Database.InfluxDB.Types as Types
 import Database.InfluxDB.Query hiding (query)
 import qualified Database.InfluxDB.Format as F
-import qualified Network.HTTP.Client.Compat as HC
+
+-- $setup
+-- >>> :set -XOverloadedStrings
+-- >>> import Database.InfluxDB.Query
+-- >>> import Database.InfluxDB.Format ((%))
 
 -- | Send a database management query to InfluxDB.
+--
+-- >>> let db = "manage-test"
+-- >>> let p = queryParams db
+-- >>> manage p $ F.formatQuery ("CREATE DATABASE "%F.database) db
 manage :: QueryParams -> Query -> IO ()
 manage params q = do
   manager' <- either HC.newManager return $ params^.manager
-  response <- HC.httpLbs request manager'
+  response <- HC.httpLbs request manager' `catch` (throwIO . HTTPException)
   let body = HC.responseBody response
   case eitherDecode' body of
-    Left message -> do
-      throwIO $ IllformedJSON message body
+    Left message ->
+      throwIO $ UnexpectedResponse message body
     Right val -> case A.parse (parseResults (params^.precision)) val of
       A.Success (_ :: V.Vector Void) -> return ()
       A.Error message -> do
@@ -53,7 +81,7 @@ manage params q = do
         when (HT.statusIsServerError status) $
           throwIO $ ServerError message
         when (HT.statusIsClientError status) $
-          throwIO $ BadRequest message request
+          throwIO $ ClientError message request
         fail $ "BUG: " ++ message ++ " in Database.InfluxDB.Manage.manage"
   where
     request = HC.setQueryString qs $ manageRequest params
@@ -72,23 +100,25 @@ manageRequest params = HC.defaultRequest
   where
     Server {..} = params^.server
 
+-- |
+-- >>> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
 data ShowQuery = ShowQuery
-  { _qid :: !Int
-  , _queryText :: !Query
-  , _database :: !Database
-  , _duration :: !NominalDiffTime
-  } deriving Show
+  { showQueryQid :: !Int
+  , showQueryText :: !Query
+  , showQueryDatabase :: !Database
+  , showQueryDuration :: !NominalDiffTime
+  }
 
 instance QueryResults ShowQuery where
   parseResults _ = parseResultsWith $ \_ _ columns fields ->
     maybe (fail "parseResults: parse error") return $ do
-      Number (toBoundedInteger -> Just _qid) <-
+      Number (toBoundedInteger -> Just showQueryQid) <-
         V.elemIndex "qid" columns >>= V.indexM fields
-      String (F.formatQuery F.text -> _queryText) <-
+      String (F.formatQuery F.text -> showQueryText) <-
         V.elemIndex "query" columns >>= V.indexM fields
-      String (F.formatDatabase F.text -> _database) <-
+      String (F.formatDatabase F.text -> showQueryDatabase) <-
         V.elemIndex "database" columns >>= V.indexM fields
-      String (parseDuration -> Right _duration) <-
+      String (parseDuration -> Right showQueryDuration) <-
         V.elemIndex "duration" columns >>= V.indexM fields
       return ShowQuery {..}
 
@@ -110,49 +140,57 @@ parseDuration = AT.parseOnly $ sum <$!> durations
 
 newtype ShowSeries = ShowSeries
   { _key :: Key
-  } deriving Show
+  }
 
 instance QueryResults ShowSeries where
-  parseResults _ = parseResultsWith $ \_ _ columns fields ->
-    ShowSeries <$> parseKey "key" columns fields
+  parseResults _ = parseResultsWith $ \_ _ columns fields -> do
+    FieldString name <- getField "key" columns fields >>= parseQueryField
+    return $ ShowSeries $ F.formatKey F.text name
 
-makeLensesWith (lensRules & generateSignatures .~ False) ''ShowQuery
+makeLensesWith
+  ( lensRules
+    & generateSignatures .~ False
+    & lensField .~ lookingupNamer
+      [ ("showQueryQid", "qid")
+      , ("showQueryText", "queryText")
+      , ("showQueryDatabase", "_database")
+      , ("showQueryDuration", "duration")
+      ]
+  ) ''ShowQuery
 
 -- | Query ID
 --
--- >>> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
--- >>> v ^.. each.qid
--- [149250]
+-- >> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
+-- >> v ^.. each.qid
+-- >[149250]
 qid :: Lens' ShowQuery Int
 
 -- | Query text
 --
--- >>> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
--- >>> v ^.. each.queryText
--- ["SHOW QUERIES"]
+-- >> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
+-- >> v ^.. each.queryText
+-- >["SHOW QUERIES"]
 queryText :: Lens' ShowQuery Query
 
-database :: Lens' ShowQuery Database
-
 -- |
--- >>> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
--- >>> v ^.. each.database
--- ["_internal"]
+-- >> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
+-- >> v ^.. each.database
+-- >["_internal"]
 instance HasDatabase ShowQuery where
-  database = Database.InfluxDB.Manage.database
+  database = _database
 
 -- | Duration of the query
 --
--- >>> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
--- >>> v ^.. each.duration
--- [0.06062s]
+-- >> v <- query (queryParams "_internal") "SHOW QUERIES" :: IO (V.Vector ShowQuery)
+-- >> v ^.. each.duration
+-- >[0.06062s]
 duration :: Lens' ShowQuery NominalDiffTime
 
 makeLensesWith (lensRules & generateSignatures .~ False) ''ShowSeries
 
 -- | Series name
 --
--- >>> v <- query (queryParams "_internal") "SHOW SERIES" :: IO (V.Vector ShowSeries)
--- >>> length $ v ^.. each.key
--- 755
+-- >> v <- query (queryParams "_internal") "SHOW SERIES" :: IO (V.Vector ShowSeries)
+-- >> length $ v ^.. each.key
+-- >755
 key :: Lens' ShowSeries Key