+++ src/Database/InfluxDB/Write.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
@@ -6,8 +7,14 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE ViewPatterns #-}
+#if __GLASGOW_HASKELL__ >= 800
+{-# OPTIONS_GHC -Wno-missing-signatures #-}
+#else
+{-# OPTIONS_GHC -fno-warn-missing-signatures #-}
+#endif
 module Database.InfluxDB.Write
   ( -- * Writers
+    -- $intro
     write
   , writeBatch
   , writeByteString
@@ -31,27 +38,43 @@ import qualified Data.Aeson.Types as A
 import qualified Data.ByteString.Char8 as B8
 import qualified Data.ByteString.Lazy as BL
 import qualified Data.Text.Encoding as TE
+import qualified Network.HTTP.Client as HC
 import qualified Network.HTTP.Types as HT
 
 import Database.InfluxDB.Line
 import Database.InfluxDB.Types as Types
 import Database.InfluxDB.JSON
-import qualified Network.HTTP.Client.Compat as HC
+
+-- $setup
+-- >>> :set -XOverloadedStrings
+-- >>> import qualified Data.Map as Map
+-- >>> import Data.Time
+-- >>> import Database.InfluxDB
+-- >>> manage (queryParams "test-db") "CREATE DATABASE \"test-db\""
+
+{- $intro
+The code snippets in this module assume the following imports.
+
+@
+import qualified Data.Map as Map
+import Data.Time
+@
+-}
 
 -- | The full set of parameters for the HTTP writer.
 data WriteParams = WriteParams
-  { _server :: !Server
-  , _database :: !Database
+  { writeServer :: !Server
+  , writeDatabase :: !Database
   -- ^ Database to be written
-  , _retentionPolicy :: !(Maybe Key)
+  , writeRetentionPolicy :: !(Maybe Key)
   -- ^ 'Nothing' means the default retention policy for the database.
-  , _precision :: !(Precision 'WriteRequest)
+  , writePrecision :: !(Precision 'WriteRequest)
   -- ^ Timestamp precision
   --
   -- In the HTTP API, timestamps are scaled by the given precision.
-  , _authentication  :: !(Maybe Credentials)
+  , writeAuthentication  :: !(Maybe Credentials)
   -- ^ No authentication by default
-  , _manager :: !(Either HC.ManagerSettings HC.Manager)
+  , writeManager :: !(Either HC.ManagerSettings HC.Manager)
   -- ^ HTTP connection manager
   }
 
@@ -59,45 +82,56 @@ data WriteParams = WriteParams
 --
 -- Default parameters:
 --
---   ['L.server'] 'localServer'
+--   ['L.server'] 'defaultServer'
 --   ['L.precision'] 'Nanosecond'
 --   ['retentionPolicy'] 'Nothing'
 --   ['L.manager'] @'Left' 'HC.defaultManagerSettings'@
 writeParams :: Database -> WriteParams
-writeParams _database = WriteParams
-  { _server = localServer
-  , _precision = Nanosecond
-  , _retentionPolicy = Nothing
-  , _authentication = Nothing
-  , _manager = Left HC.defaultManagerSettings
+writeParams writeDatabase = WriteParams
+  { writeServer = defaultServer
+  , writePrecision = Nanosecond
+  , writeRetentionPolicy = Nothing
+  , writeAuthentication = Nothing
+  , writeManager = Left HC.defaultManagerSettings
   , ..
   }
 
--- | Write a 'Line'
+-- | Write a 'Line'.
+--
+-- >>> let p = writeParams "test-db"
+-- >>> write p $ Line "room_temp" Map.empty (Map.fromList [("temp", FieldFloat 25.0)]) (Nothing :: Maybe UTCTime)
 write
   :: Timestamp time
   => WriteParams
   -> Line time
   -> IO ()
-write p@WriteParams {_precision} =
-  writeByteString p . encodeLine (scaleTo _precision)
+write p@WriteParams {writePrecision} =
+  writeByteString p . encodeLine (scaleTo writePrecision)
 
--- | Write 'Line's in a batch
+-- | Write multiple 'Line's in a batch.
 --
--- This is more efficient than 'write'.
+-- This is more efficient than calling 'write' multiple times.
+--
+-- >>> let p = writeParams "test-db"
+-- >>> :{
+-- writeBatch p
+--   [ Line "temp" (Map.singleton "city" "tokyo") (Map.fromList [("temp", FieldFloat 25.0)]) (Nothing :: Maybe UTCTime)
+--   , Line "temp" (Map.singleton "city" "osaka") (Map.fromList [("temp", FieldFloat 25.2)]) (Nothing :: Maybe UTCTime)
+--   ]
+-- :}
 writeBatch
   :: (Timestamp time, Foldable f)
   => WriteParams
   -> f (Line time)
   -> IO ()
-writeBatch p@WriteParams {_precision} =
-  writeByteString p . encodeLines (scaleTo _precision)
+writeBatch p@WriteParams {writePrecision} =
+  writeByteString p . encodeLines (scaleTo writePrecision)
 
 -- | Write a raw 'BL.ByteString'
 writeByteString :: WriteParams -> BL.ByteString -> IO ()
 writeByteString params payload = do
-  manager' <- either HC.newManager return $ _manager params
-  response <- HC.httpLbs request manager'
+  manager' <- either HC.newManager return $ writeManager params
+  response <- HC.httpLbs request manager' `catch` (throwIO . HTTPException)
   let body = HC.responseBody response
       status = HC.responseStatus response
   if BL.null body
@@ -106,18 +140,19 @@ writeByteString params payload = do
       when (HT.statusIsServerError status) $
         throwIO $ ServerError message
       when (HT.statusIsClientError status) $
-        throwIO $ BadRequest message request
+        throwIO $ ClientError message request
     else case A.eitherDecode' body of
       Left message ->
-        throwIO $ IllformedJSON message body
+        throwIO $ UnexpectedResponse message body
       Right val -> case A.parse parseErrorObject val of
         A.Success _ ->
-          fail $ "BUG: impossible code path in Database.InfluxDB.Write.writeByteString"
+          fail $ "BUG: impossible code path in "
+            ++ "Database.InfluxDB.Write.writeByteString"
         A.Error message -> do
           when (HT.statusIsServerError status) $
             throwIO $ ServerError message
           when (HT.statusIsClientError status) $
-            throwIO $ BadRequest message request
+            throwIO $ ClientError message request
           fail $ "BUG: " ++ message
             ++ " in Database.InfluxDB.Write.writeByteString"
 
@@ -136,74 +171,77 @@ writeRequest WriteParams {..} =
     , HC.path = "/write"
     }
   where
-    Server {..} = _server
+    Server {..} = writeServer
     qs = concat
-      [ [("db", Just $ TE.encodeUtf8 $ databaseName _database)]
+      [ [("db", Just $ TE.encodeUtf8 $ databaseName writeDatabase)]
       , fromMaybe [] $ do
-        Key name <- _retentionPolicy
+        Key name <- writeRetentionPolicy
         return [("rp", Just (TE.encodeUtf8 name))]
       , fromMaybe [] $ do
-        Credentials { _user = u, _password = p } <- _authentication
+        Credentials { _user = u, _password = p } <- writeAuthentication
         return
           [ ("u", Just (TE.encodeUtf8 u))
           , ("p", Just (TE.encodeUtf8 p))
           ]
       ]
 
-makeLensesWith (lensRules & generateSignatures .~ False) ''WriteParams
-
-server :: Lens' WriteParams Server
+makeLensesWith
+  ( lensRules
+    & generateSignatures .~ False
+    & lensField .~ lookingupNamer
+      [ ("writeServer", "_server")
+      , ("writeDatabase", "_database")
+      , ("writeRetentionPolicy", "retentionPolicy")
+      , ("writePrecision", "_precision")
+      , ("writeManager", "_manager")
+      , ("writeAuthentication", "_authentication")
+      ]
+    )
+  ''WriteParams
 
 -- |
 -- >>> let p = writeParams "foo"
 -- >>> p ^. server.host
 -- "localhost"
 instance HasServer WriteParams where
-  server = Database.InfluxDB.Write.server
-
-database :: Lens' WriteParams Database
+  server = _server
 
 -- |
 -- >>> let p = writeParams "foo"
 -- >>> p ^. database
 -- "foo"
 instance HasDatabase WriteParams where
-  database = Database.InfluxDB.Write.database
+  database = _database
 
 -- | Target retention policy for the write.
 --
 -- InfluxDB writes to the @default@ retention policy if this parameter is set
 -- to 'Nothing'.
 --
--- >>> let p = writeParams "foo"
--- >>> let p' = p & retentionPolicy .~ Just "two_hours"
--- >>> p' ^. retentionPolicy
+-- >>> let p = writeParams "foo" & retentionPolicy .~ Just "two_hours"
+-- >>> p ^. retentionPolicy
 -- Just "two_hours"
 retentionPolicy :: Lens' WriteParams (Maybe Key)
-
-precision :: Lens' WriteParams (Precision 'WriteRequest)
 
 -- |
 -- >>> let p = writeParams "foo"
 -- >>> p ^. precision
 -- Nanosecond
 instance HasPrecision 'WriteRequest WriteParams where
-  precision = Database.InfluxDB.Write.precision
-
-manager :: Lens' WriteParams (Either HC.ManagerSettings HC.Manager)
+  precision = _precision
 
 -- |
--- >>> let p = writeParams "foo"
--- >>> p & manager .~ Left HC.defaultManagerSettings
+-- >>> let p = writeParams "foo" & manager .~ Left HC.defaultManagerSettings
 instance HasManager WriteParams where
-  manager = Database.InfluxDB.Write.manager
+  manager = _manager
 
 -- | Authentication info for the write
 --
 -- >>> let p = writeParams "foo"
 -- >>> p ^. authentication
 -- Nothing
-authentication :: Lens' WriteParams (Maybe Credentials)
-
+-- >>> let p' = p & authentication ?~ credentials "john" "passw0rd"
+-- >>> p' ^. authentication . traverse . user
+-- "john"
 instance HasCredentials WriteParams where
-  authentication = Database.InfluxDB.Write.authentication
+  authentication = _authentication